api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi>=0.127.0",
    "google-search-results>=2.4.2",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain[google-genai]>=1.2.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "serpapi>=0.1.5",
    "supabase>=2.27.0",
    "uvicorn[standard]>=0.40.0",
]
